Django-Backend/requirements.txt
django
djangorestframework
django-cors-headers
django-allauth
django-registration
django-pwa
django-extensions